neuronske_mreze/patterns_cnn/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.1.0=gpu
absl-py=0.9.0=py36_0
alabaster=0.7.12=py36_0
argh=0.26.2=py36_0
astor=0.8.0=py36_0
astroid=2.4.0=py36_0
atomicwrites=1.4.0=py_0
attrs=19.3.0=py_0
automat=0.8.0=py_0
autopep8=1.4.4=py_0
babel=2.8.0=py_0
backcall=0.1.0=py36_0
bcrypt=3.1.7=py36he774522_0
blas=1.0=mkl
bleach=3.1.4=py_0
blinker=1.4=py36_0
ca-certificates=2020.1.1=0
cachetools=3.1.1=py_0
certifi=2020.4.5.1=py36_0
cffi=1.14.0=py36h7a1dbc1_0
chardet=3.0.4=py36_1003
click=7.1.2=py_0
cloudpickle=1.4.1=py_0
colorama=0.4.3=py_0
constantly=15.1.0=py36h28b3542_0
cryptography=2.9.2=py36h7a1dbc1_0
cssselect=1.1.0=py_0
cudatoolkit=10.1.243=h74a9793_0
cudnn=7.6.5=cuda10.1_0
cycler=0.10.0=py36h009560c_0
cytoolz=0.10.1=py36he774522_0
dask-core=2.16.0=py_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
diff-match-patch=20181111=py_0
docutils=0.16=py36_0
entrypoints=0.3=py36_0
flake8=3.7.9=py36_0
freetype=2.9.1=ha9979f8_1
future=0.18.2=py36_0
gast=0.2.2=py36_0
google-auth=1.14.1=py_0
google-auth-oauthlib=0.4.1=py_2
google-pasta=0.2.0=py_0
grpcio=1.27.2=py36h351948d_0
h5py=2.10.0=py36h5e291fa_0
hdf5=1.10.4=h7ebc959_0
hyperlink=19.0.0=py_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=2.9=py_1
imageio=2.8.0=py_0
imagesize=1.2.0=py_0
importlib_metadata=1.5.0=py36_0
incremental=17.5.0=py36_0
intel-openmp=2020.1=216
intervaltree=3.0.2=py_0
ipykernel=5.1.4=py36h39e3cac_0
ipython=7.13.0=py36h5ca1d4c_0
ipython_genutils=0.2.0=py36_0
isort=4.3.21=py36_0
jedi=0.15.2=py36_0
jinja2=2.11.2=py_0
joblib=0.14.1=py_0
jpeg=9b=hb83a4c4_2
jsonschema=3.2.0=py36_0
jupyter_client=6.1.3=py_0
jupyter_core=4.6.3=py36_0
keras=2.3.1=0
keras-applications=1.0.8=py_0
keras-base=2.3.1=py36_0
keras-preprocessing=1.1.0=py_1
keyring=21.1.1=py36_2
kiwisolver=1.2.0=py36h74a9793_0
lazy-object-proxy=1.4.3=py36he774522_0
libiconv=1.15=h1df5818_7
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.11.4=h7bd577a_0
libsodium=1.0.16=h9d3ae62_0
libspatialindex=1.9.3=h33f27b4_0
libtiff=4.1.0=h56a325e_0
libxml2=2.9.9=h464c3ec_0
libxslt=1.1.33=h579f668_0
lxml=4.5.0=py36h1350720_0
markdown=3.1.1=py36_0
markupsafe=1.1.1=py36he774522_0
matplotlib=3.1.3=py36_0
matplotlib-base=3.1.3=py36h64f37c6_0
mccabe=0.6.1=py36_1
mistune=0.8.4=py36he774522_0
mkl=2020.1=216
mkl-service=2.3.0=py36hb782905_0
mkl_fft=1.0.15=py36h14836fe_0
mkl_random=1.1.0=py36h675688f_0
more-itertools=8.2.0=py_0
nbconvert=5.6.1=py36_0
nbformat=5.0.6=py_0
networkx=2.4=py_0
numpy=1.18.1=py36h93ca92e_0
numpy-base=1.18.1=py36hc3f5095_1
numpydoc=0.9.2=py_0
oauthlib=3.1.0=py_0
olefile=0.46=py36_0
openssl=1.1.1g=he774522_0
opt_einsum=3.1.0=py_0
packaging=20.3=py_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py36_1
paramiko=2.7.1=py_0
parsel=1.5.2=py36_0
parso=0.5.2=py_0
pathtools=0.1.2=py_1
pexpect=4.8.0=py36_0
pickleshare=0.7.5=py36_0
pillow=7.1.2=py36hcc1f983_0
pip=20.0.2=py36_3
pluggy=0.13.1=py36_0
prompt-toolkit=3.0.4=py_0
prompt_toolkit=3.0.4=0
protobuf=3.11.4=py36h33f27b4_0
psutil=5.7.0=py36he774522_0
py=1.8.1=py_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycodestyle=2.5.0=py36_0
pycparser=2.20=py_0
pydispatcher=2.0.5=py36_1
pydocstyle=4.0.1=py_0
pyflakes=2.1.1=py36_0
pygments=2.6.1=py_0
pyhamcrest=1.9.0=py36_2
pyjwt=1.7.1=py36_0
pylint=2.5.0=py36_1
pynacl=1.3.0=py36h62dcd97_0
pyopenssl=19.1.0=py36_0
pyparsing=2.4.7=py_0
pyqt=5.9.2=py36h6538335_2
pyreadline=2.1=py36_1
pyrsistent=0.16.0=py36he774522_0
pysocks=1.7.1=py36_0
pytest=5.4.2=py36_0
pytest-runner=5.2=py_0
python=3.6.10=h9f7ef89_2
python-dateutil=2.8.1=py_0
python-jsonrpc-server=0.3.4=py_0
python-language-server=0.31.10=py36_0
pytz=2020.1=py_0
pywavelets=1.1.1=py36he774522_0
pywin32=227=py36he774522_1
pywin32-ctypes=0.2.0=py36_1000
pyyaml=5.3.1=py36he774522_0
pyzmq=18.1.1=py36ha925a31_0
qdarkstyle=2.8.1=py_0
qt=5.9.7=vc14h73c81de_0
qtawesome=0.7.0=py_0
qtconsole=4.7.4=py_0
qtpy=1.9.0=py_0
queuelib=1.5.0=py36_0
requests=2.23.0=py36_0
requests-oauthlib=1.3.0=py_0
rope=0.17.0=py_0
rsa=4.0=py_0
rtree=0.9.4=py36h21ff451_1
scikit-image=0.16.2=py36h47e9c7a_0
scikit-learn=0.22.1=py36h6288b17_0
scipy=1.4.1=py36h9439919_0
scrapy=1.6.0=py36_0
service_identity=18.1.0=py36h28b3542_0
setuptools=46.2.0=py36_0
sip=4.19.8=py36h6538335_0
six=1.14.0=py36_0
snowballstemmer=2.0.0=py_0
sortedcontainers=2.1.0=py36_0
sphinx=3.0.3=py_0
sphinxcontrib-applehelp=1.0.2=py_0
sphinxcontrib-devhelp=1.0.2=py_0
sphinxcontrib-htmlhelp=1.0.3=py_0
sphinxcontrib-jsmath=1.0.1=py_0
sphinxcontrib-qthelp=1.0.3=py_0
sphinxcontrib-serializinghtml=1.1.4=py_0
spyder=4.1.3=py36_0
spyder-kernels=1.9.1=py36_0
sqlite=3.31.1=h2a8f88b_1
tensorboard=2.1.0=py3_0
tensorflow=2.1.0=gpu_py36h3346743_0
tensorflow-base=2.1.0=gpu_py36h55f5790_0
tensorflow-estimator=2.1.0=pyhd54b08b_0
tensorflow-gpu=2.1.0=h0d30ee6_0
termcolor=1.1.0=py36_1
testpath=0.4.4=py_0
tk=8.6.8=hfa6e2cd_0
toml=0.10.0=py36h28b3542_0
toolz=0.10.0=py_0
tornado=6.0.4=py36he774522_1
traitlets=4.3.3=py36_0
twisted=20.3.0=py36he774522_0
typed-ast=1.4.1=py36he774522_0
ujson=1.35=py36hfa6e2cd_0
urllib3=1.25.8=py36_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_1
w3lib=1.21.0=py_0
watchdog=0.10.2=py36_0
wcwidth=0.1.9=py_0
webencodings=0.5.1=py36_1
werkzeug=0.14.1=py36_0
wheel=0.34.2=py36_0
win_inet_pton=1.1.0=py36_0
wincertstore=0.2=py36h7fe50ca_0
wrapt=1.11.2=py36he774522_0
xz=5.2.5=h62dcd97_0
yaml=0.1.7=hc54c509_2
yapf=0.28.0=py_0
zeromq=4.3.1=h33f27b4_3
zipp=3.1.0=py_0
zlib=1.2.11=h62dcd97_4
zope=1.0=py36_1
zope.interface=4.7.1=py36he774522_0
zstd=1.3.7=h508b16e_0
